+++ lib/src/generator/entity_type_helper.dart
@@ -4,29 +4,33 @@ import 'package:json_serializable/type_helper.dart';
 
 import 'identifier.dart';
 
+/// Helper which looks up if a [DartType] exists within a list of [_entities]
+/// and returns the code for the proper calls to encode/decode method. The list
+/// of entities is provided by the generator based on entities, transformers
+/// and replacements.
 class EntityTypeHelper implements TypeHelper {
-  final Map<ClassElement, ClassElement> entities;
-  EntityTypeHelper(this.entities);
+  final Map<ClassElement, ClassElement> _entities;
+  EntityTypeHelper(this._entities);
 
   @override
   String deserialize(
       DartType targetType, String expression, DeserializeContext context) {
     final target = targetType.element as ClassElement;
-    if (!entities.containsKey(target)) {
+    if (!_entities.containsKey(target)) {
       return null;
     }
 
-    return '${decodeMethod(entities[target])}($expression, inst)';
+    return '${decodeMethod(_entities[target])}($expression, inst)';
   }
 
   @override
   String serialize(
       DartType targetType, String expression, SerializeContext context) {
     final target = targetType.element as ClassElement;
-    if (!entities.containsKey(target)) {
+    if (!_entities.containsKey(target)) {
       return null;
     }
 
-    return '${encodeMethod(entities[target])}($expression, inst)';
+    return '${encodeMethod(_entities[target])}($expression, inst)';
   }
 }