+++ test/src/serializer.g.dart
@@ -126,15 +126,9 @@ MyImpl _MyImpl$decoder(Map<String, dynamic> data, Dartson inst) {
   return obj;
 }
 
-class _Dartson$impl extends Dartson<Map<String, dynamic>> {
-  _Dartson$impl()
-      : super(<Type, DartsonEntity>{
-          MyClass:
-              const DartsonEntity<MyClass>(_MyClass$encoder, _MyClass$decoder),
-          SubClass: const DartsonEntity<SubClass>(
-              _SubClass$encoder, _SubClass$decoder),
-          MyImpl: const DartsonEntity<MyImpl>(_MyImpl$encoder, _MyImpl$decoder)
-        });
-}
-
-final _serializer$dartson = new _Dartson$impl();
+final _serializer$dartson =
+    new Dartson<Map<String, dynamic>>(<Type, DartsonEntity>{
+  MyClass: const DartsonEntity<MyClass>(_MyClass$encoder, _MyClass$decoder),
+  SubClass: const DartsonEntity<SubClass>(_SubClass$encoder, _SubClass$decoder),
+  MyImpl: const DartsonEntity<MyImpl>(_MyImpl$encoder, _MyImpl$decoder)
+});